refactor(lifecycle): remove settle phase, inline into lifecycle
