--- src/lifecycle-settle.ts
@@ -1,46 +1,0 @@
-import { type RecoveryAction, recoveryActionForError as resolveRecoveryAction } from "./error-handling";
-import { t } from "./i18n";
-import type { LifecycleInput, RunContext } from "./lifecycle-contract";
-import { defaultLifecyclePolicy, type LifecyclePolicy } from "./lifecycle-policy";
-import { acceptedLifecycleSignal } from "./lifecycle-state";
-
-export function recoveryActionForError(
-  input: { errorCode?: string; unknownErrorCount: number },
-  policy: LifecyclePolicy = defaultLifecyclePolicy,
-): RecoveryAction {
-  return resolveRecoveryAction(input, policy.maxUnknownErrorsPerRequest);
-}
-
-export async function phaseSettle(ctx: RunContext, shouldYield: LifecycleInput["shouldYield"]) {
-  if (!ctx.result) return;
-  if (shouldYield?.()) return;
-
-  const lifecycleSignal = acceptedLifecycleSignal(ctx);
-  if (lifecycleSignal) {
-    ctx.currentError = undefined;
-    ctx.debug("lifecycle.signal.accepted", {
-      signal: lifecycleSignal,
-      tool_calls: ctx.result.toolCalls.length,
-    });
-  }
-
-  if (
-    recoveryActionForError({
-      errorCode: ctx.currentError?.code,
-      unknownErrorCount: ctx.errorStats.other,
-    }) === "stop-unknown-budget"
-  ) {
-    ctx.debug("lifecycle.eval.skipped", {
-      reason: "unknown_error_budget",
-      unknown_error_count: ctx.errorStats.other,
-      unknown_error_cap: ctx.policy.maxUnknownErrorsPerRequest,
-      last_error_code: ctx.currentError?.code ?? null,
-    });
-    if (!ctx.result.text.trim()) {
-      ctx.result = {
-        text: t("lifecycle.stopped_unknown_errors"),
-        toolCalls: [],
-      };
-    }
-  }
-}
